requirements.txt
selenium==4.27.1
webdriver-manager==4.0.2
gspread==6.1.4
oauth2client==4.1.3
python-dotenv==1.0.1
pandas==2.2.3
FreeSimpleGUI==5.2.0
APScheduler==3.11.0
pystray==0.19.5